Added E alphabet detectors pattern test cases
