+++ pkg/detectors/endorlabs/endorlabs_test.go
@@ -2,10 +2,11 @@ package endorlabs
 
 import (
 	"context"
+	"testing"
+
 	"github.com/google/go-cmp/cmp"
 	"github.com/trufflesecurity/trufflehog/v3/pkg/detectors"
 	"github.com/trufflesecurity/trufflehog/v3/pkg/engine/ahocorasick"
-	"testing"
 )
 
 func TestEndorlabs_Pattern(t *testing.T) {
