+++ pkg/detectors/eraser/eraser_test.go
@@ -2,10 +2,12 @@ package eraser
 
 import (
 	"context"
+	"testing"
+
 	"github.com/google/go-cmp/cmp"
+
 	"github.com/trufflesecurity/trufflehog/v3/pkg/detectors"
 	"github.com/trufflesecurity/trufflehog/v3/pkg/engine/ahocorasick"
-	"testing"
 )
 
 func TestEraser_Pattern(t *testing.T) {
